<SearchPageNext /> -> <SearchPage />, with upgrade instructions and deprecation warnings.
