+++ plugins/search/src/components/index.tsx
@@ -19,7 +19,6 @@ export * from './SearchFilterNext';
 export * from './SearchBar';
 export * from './SearchBarNext';
 export * from './SearchPage';
-export * from './SearchPageNext';
 export * from './SearchResult';
 export * from './SearchResultNext';
 export * from './DefaultResultListItem';
